Finished the majority of the infinite terrain generation.
The infinite terrain generation now has a LOD system  so that chunks further away have fewer vertices in a chunk. The different distances at which this occurs is changable through the scene's "MapGeneration" game object in the editor.

The infinite terrain now has a colour map texture for each chunk.

Each chunk now has its mesh generated acording to the noise map offset from the center chunk, although the seams between each chunk are not perfect and often have gaps (this is a bug which will be fixed in a later commit).