Added a falloff map to the list of generation options, also started work on fixing triangle texture lighting problems.
A falloff map option has been added to the map generator to allow the ability of generating only islands.

I started work on fixing a mesh triangle lighting bug that happens on the edge of chunks. It happens because the triangles on the edge of chunks only used the normals of triangles from the same chunk for its lighting calculations, thereby missing the information of the triangles next to it because they are in two different chunks.