Updated alot.
Added collions to the generated chunk meshes, which allows for a rigidbody to move on the generated terrain.

Added the option for flatshading.
This comes at the cost of a smaller chunk size, to account for the extra vertices needed for the shading.

Changed the way the data storage worked. It now uses unique assets with  their own values, to change the noise and terrain generation. This will allow the saving of those assets, for different combinations of noise and terrain generation.

Also fixed a chunk size bug that happened with the falloff map generation.